Title: 낮은 지점들, Time: 469ms, Memory: 24MB, Status: Passed - Codetree

+++ 250810/낮은 지점들/lowest-points.java
@@ -23,7 +23,7 @@ public class Main {
             }
         }
 
-        int answer = 0;
+        Long answer = 0L;
         for (int v : m.keySet()) {
             answer += m.get(v);
         }
